feat: move note/types composables to commons (batch 17)
Move EcashMint, CodeSnippet, NIP90Status composables to commons/commonMain.
Add GenericLoadable, BookmarkType, subtleBorder, replyModifier to commons.
Cherry-pick Route/INav/IAccountViewModel foundation from architecture-unblocks.

Files moved:
- EcashMint.kt (RenderCashuMint, RenderFedimint, RenderMintRecommendation)
- CodeSnippet.kt (RenderCodeSnippetEvent, RenderCodeSnippetHeaderForThread)
- NIP90Status.kt (RenderNIP90Status)

Supporting moves:
- GenericLoadable sealed class → commons/ui/components/
- BookmarkType enum → commons/ui/screen/loggedIn/bookmarkgroups/
- subtleBorder, replyModifier extensions → commons/ui/theme/ThemeExtensions.kt

Part of the KMP iOS migration tracked in #2238.
