+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/types/CodeSnippet.kt
@@ -18,225 +18,23 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+@file:Suppress("unused")
+
 package com.vitorpamplona.amethyst.ui.note.types
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
+// Re-export from commons for backwards compatibility
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.theme.QuoteBorder
-import com.vitorpamplona.amethyst.ui.theme.StdHorzSpacer
-import com.vitorpamplona.amethyst.ui.theme.StdVertSpacer
-import com.vitorpamplona.amethyst.ui.theme.replyModifier
 import com.vitorpamplona.quartz.nipC0CodeSnippets.CodeSnippetEvent
+import com.vitorpamplona.amethyst.commons.ui.note.types.RenderCodeSnippetEvent as CommonsRenderCodeSnippetEvent
+import com.vitorpamplona.amethyst.commons.ui.note.types.RenderCodeSnippetHeaderForThread as CommonsRenderCodeSnippetHeaderForThread
 
 @Composable
 fun RenderCodeSnippetEvent(baseNote: Note) {
-    val noteEvent = baseNote.event as? CodeSnippetEvent ?: return
-
-    val name =
-        remember(noteEvent) {
-            noteEvent.snippetName()
-                ?: noteEvent.language()?.let { "$it snippet" }
-                ?: "Code Snippet"
-        }
-    val language = remember(noteEvent) { noteEvent.language() ?: noteEvent.extension() }
-    val description = remember(noteEvent) { noteEvent.snippetDescription() }
-    val codePreview =
-        remember(noteEvent) {
-            noteEvent.content
-                .lines()
-                .take(5)
-                .joinToString("\n")
-        }
-
-    Column(MaterialTheme.colorScheme.replyModifier) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 10.dp, vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = name,
-                style = MaterialTheme.typography.titleMedium,
-                modifier = Modifier.weight(1f),
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            language?.let {
-                Spacer(modifier = StdHorzSpacer)
-                Box(
-                    modifier =
-                        Modifier
-                            .clip(QuoteBorder)
-                            .background(MaterialTheme.colorScheme.surfaceVariant)
-                            .padding(horizontal = 6.dp, vertical = 2.dp),
-                ) {
-                    Text(
-                        text = it,
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-            }
-        }
-
-        description?.ifBlank { null }?.let {
-            Text(
-                text = it,
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 10.dp),
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-            )
-            Spacer(modifier = StdVertSpacer)
-        }
-
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(10.dp)
-                    .clip(QuoteBorder)
-                    .background(MaterialTheme.colorScheme.onSurface.copy(alpha = 0.05f))
-                    .border(1.dp, MaterialTheme.colorScheme.outline.copy(alpha = 0.3f), QuoteBorder)
-                    .padding(10.dp),
-        ) {
-            Text(
-                text = codePreview,
-                style =
-                    TextStyle(
-                        fontFamily = FontFamily.Monospace,
-                        fontSize = 12.sp,
-                    ),
-                maxLines = 5,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
+    CommonsRenderCodeSnippetEvent(baseNote)
 }
 
 @Composable
 fun RenderCodeSnippetHeaderForThread(noteEvent: CodeSnippetEvent) {
-    val name =
-        remember(noteEvent) {
-            noteEvent.snippetName()
-                ?: noteEvent.language()?.let { "$it snippet" }
-                ?: "Code Snippet"
-        }
-    val language = remember(noteEvent) { noteEvent.language() ?: noteEvent.extension() }
-    val description = remember(noteEvent) { noteEvent.snippetDescription() }
-    val license = remember(noteEvent) { noteEvent.license() }
-    val runtime = remember(noteEvent) { noteEvent.runtime() }
-    val deps = remember(noteEvent) { noteEvent.deps() }
-
-    Column(modifier = Modifier.padding(start = 12.dp, end = 12.dp, bottom = 12.dp)) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = name,
-                style = MaterialTheme.typography.headlineSmall,
-                modifier = Modifier.weight(1f),
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            language?.let {
-                Spacer(modifier = StdHorzSpacer)
-                Box(
-                    modifier =
-                        Modifier
-                            .clip(QuoteBorder)
-                            .background(MaterialTheme.colorScheme.surfaceVariant)
-                            .padding(horizontal = 8.dp, vertical = 3.dp),
-                ) {
-                    Text(
-                        text = it,
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-            }
-        }
-
-        description?.ifBlank { null }?.let {
-            Spacer(modifier = StdVertSpacer)
-            Text(
-                text = it,
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        }
-
-        Spacer(modifier = StdVertSpacer)
-
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .clip(QuoteBorder)
-                    .background(MaterialTheme.colorScheme.onSurface.copy(alpha = 0.05f))
-                    .border(1.dp, MaterialTheme.colorScheme.outline.copy(alpha = 0.3f), QuoteBorder)
-                    .padding(12.dp),
-        ) {
-            Text(
-                text = noteEvent.content,
-                style =
-                    TextStyle(
-                        fontFamily = FontFamily.Monospace,
-                        fontSize = 13.sp,
-                    ),
-            )
-        }
-
-        license?.let {
-            Spacer(modifier = StdVertSpacer)
-            Text(
-                text = "License: $it",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-
-        runtime?.let {
-            Text(
-                text = "Runtime: $it",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-
-        if (deps.isNotEmpty()) {
-            Text(
-                text = "Deps: ${deps.joinToString(", ")}",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-    }
+    CommonsRenderCodeSnippetHeaderForThread(noteEvent)
 }
